+++ lib/basic_http_interceptor.dart
@@ -3,10 +3,19 @@
 /// More dartdocs go here.
 library basic_http_interceptor;
 
-import 'package:logging/logging.dart';
+import 'dart:convert';
 
+import 'package:http_interceptor/extensions/streamed_response.dart';
 import 'package:http_interceptor/http_interceptor.dart'
-    show BaseRequest, BaseResponse, InterceptorContract, Request, Response;
+    show
+        BaseRequest,
+        BaseResponse,
+        InterceptorContract,
+        Request,
+        Response,
+        StreamedResponse;
+
+import 'package:logging/logging.dart';
 
 export 'package:http_interceptor/utils/query_parameters.dart'
     show buildUrlString;
@@ -17,7 +26,6 @@ export 'utils/basic_http_interceptor_webclient.dart'
 part 'src/basic_http_interceptor_logger.dart';
 part 'src/basic_http_interceptor_header.dart';
 part 'src/basic_http_interceptor_hook.dart';
-
 
 // export 'src/basic_http_interceptor_base.dart';